+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - [#2](https://github.com/devatherock/ldap-search-api/issues/2): Unit tests
+- [#9](https://github.com/devatherock/ldap-search-api/issues/9): Integration tests
 
 ### Changed
 - Upgraded micronaut to `3.9.1`